+++ admin/client/utils/queryParams.js
@@ -1,0 +1,47 @@
+import assign from 'object-assign';
+
+/**
+ * Updates the query parameters with the ones passed as the first argument
+ *
+ * @param  {Object} params         The new parameters to be added
+ * @param  {Object} location       The current location object
+ */
+export function updateQueryParams (params, location) {
+	if (!location) return;
+	const newParams = assign({}, location.query);
+	// Stringify nested objects inside the parameters
+	Object.keys(params).forEach(i => {
+		if (params[i]) {
+			newParams[i] = params[i];
+			if (typeof newParams[i] === 'object') {
+				newParams[i] = JSON.stringify(newParams[i]);
+			}
+		} else {
+			delete newParams[i];
+		}
+	});
+
+	return newParams;
+}
+
+/**
+ * Stringify the columns array from the state
+ *
+ * @param  {Array}  columns            The columns from the active state
+ * @param  {String} defaultColumnPaths The default column paths of the current list
+ *
+ * @return {String}                    The column array, stringified
+ */
+export function stringifyColumns (columns, defaultColumnPaths) {
+	if (!columns) {
+		return;
+	}
+	// Turns [{ path: 'someColumn' }, { path: 'someOtherColumn' }]
+	// into ['someColumn', 'someOtherColumn']
+	let columnString = columns.map((column) => column.path);
+	// Turns that array into 'someColumn,someOtherColumn'
+	if (Array.isArray(columnString)) columnString = columnString.join(',');
+	// If that is the same as the default columns, don't set the query param
+	if (columnString === defaultColumnPaths) columnString = undefined;
+	return columnString;
+}
